requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
tensorflow==2.12.0
pillow==9.5.0
numpy==1.23.5